Rename op_codes to program
